+++ pyespn/classes/draft.py
@@ -32,7 +32,7 @@ class DraftPick:
         """
 
         self.pick_json = pick_json
-        self.espn_instance = espn_instance
+        self._espn_instance = espn_instance
         self.athlete = None
         self.team = None
         self._get_pick_data()
@@ -55,12 +55,18 @@ class DraftPick:
         self.overall_number = self.pick_json.get('overall')
         team_id = get_team_id(self.pick_json.get('team', {}).get('$ref'))
         athlete_url = self.pick_json.get('athlete', {}).get('$ref')
-        self.team = self.espn_instance.get_team_by_id(team_id=team_id)
+        self.team = self._espn_instance.get_team_by_id(team_id=team_id)
 
         athlete_content = fetch_espn_data(athlete_url)
 
         self.athlete = Player(player_json=athlete_content,
                               espn_instance=self.espn_instance)
+    @property
+    def espn_instance(self):
+        """
+            PYESPN: the espn client instance associated with the class
+        """
+        return self._espn_instance
 
     def to_dict(self) -> dict:
         """