+++ pyespn/classes/vehicle.py
@@ -30,7 +30,7 @@ class Vehicle:
             espn_instance (PYESPN): An instance of the ESPN API client.
         """
         self.vehicle_json = vehicle_json
-        self.espn_instance = espn_instance
+        self._espn_instance = espn_instance
         self._set_vehicle_data()
 
     def __repr__(self) -> str:
@@ -52,6 +52,13 @@ class Vehicle:
         self.engine = self.vehicle_json.get('engine')
         self.tire = self.vehicle_json.get('tire')
         self.team = self.vehicle_json.get('team')
+
+    @property
+    def espn_instance(self):
+        """
+            PYESPN: the espn client instance associated with the class
+        """
+        return self._espn_instance
 
     def to_dict(self) -> dict:
         """